Delete project resource (#423)
* Offer to delete resources in CLI and Web

* Allow delete/restore deleting project

* Fix unit test

* Send notification after delete from engine

* Bugfixes

* Fix feedback

* Remove RemoveResources option in remove project

+++ src/CLI/Polyrific.Catapult.Cli/Commands/Project/RemoveCommand.cs
@@ -12,10 +12,13 @@ namespace Polyrific.Catapult.Cli.Commands.Project
     public class RemoveCommand : BaseCommand
     {
         private readonly IProjectService _projectService;
+        private readonly IJobDefinitionService _jobDefinitionService;
 
-        public RemoveCommand(IConsole console, ILogger<RemoveCommand> logger, IProjectService projectService) : base(console, logger)
+        public RemoveCommand(IConsole console, ILogger<RemoveCommand> logger, 
+            IProjectService projectService, IJobDefinitionService jobDefinitionService) : base(console, logger)
         {
             _projectService = projectService;
+            _jobDefinitionService = jobDefinitionService;
         }
 
         [Required]
@@ -37,9 +40,19 @@ namespace Polyrific.Catapult.Cli.Commands.Project
             var project = _projectService.GetProjectByName(Name).Result;
             if (project != null)
             {
-                _projectService.DeleteProject(project.Id).Wait();
+                var deletionJobDefinition = _jobDefinitionService.GetDeletionJobDefinition(project.Id).Result;
+                if (deletionJobDefinition != null && 
+                    (AutoConfirm || Console.GetYesNo("Do you want to remove the related resources as well?", false)))
+                {
+                    _projectService.MarkProjectDeleting(project.Id).Wait();
+                    message = $"Project {Name} is being removed. You will be notified once the process has been done";
+                }
+                else
+                {
+                    _projectService.DeleteProject(project.Id).Wait();
+                    message = $"Project {Name} has been removed successfully";
+                }
 
-                message = $"Project {Name} has been removed successfully";
                 Logger.LogInformation(message);
             }
             else
